Move Handover to common module
